feat:【数据统计】用户态数据统计- 总览 #6587

+++ src/backend/ci/core/metrics/api-metrics/src/main/kotlin/com/tencent/devops/metrics/pojo/message/TurboReportEvent.kt
@@ -37,5 +37,5 @@ data class TurboReportEvent(
     @ApiModelProperty("项目ID")
     val projectId: String,
     @ApiModelProperty("编译加速节省时间，单位：秒")
-    val turboSaveTime: Long
+    val turboSaveTime: Double
 )
